+++ packages/rest/src/session/ConnectionPropsForSessCfg.ts
@@ -287,6 +287,9 @@ export class ConnectionPropsForSessCfg {
     /**
      * List of properties on `sessCfg` object that should be kept secret and
      * may not appear in Imperative log files.
+     *
+     * NOTE(Kelosky): redundant from LoggerUtils.SECURE_PROMPT_OPTIONS - leaving
+     * for future date to consolidate
      */
     private static secureSessCfgProps: Set<string> = new Set(["user", "password", "tokenValue", "passphrase"]);